Fix communication with LS240 and curve upload
The maximum number of breakpoints for a calibration curve on a given channel is
200. The assertion in load_curve() did not allow for the 200th breakpoint to be
uploaded.

When uploading a full 200 point curve I noticed the last ~40 points or so
weren't getting uploaded. I would often note the device timing out when
investigating the uploaded curves after upload. This led me to putting a 0.1
second sleep between each load_curve_point() call, which allowed the full curve
to be uploaded. Looking in the Lakeshore 240 manual, they require that 10 ms
between every command. It looks like this was implemented in the past, but was
commented out. Remove the comment, and force a 10 ms pause after every
LS240.msg() call.