Determine units for get_reading from Channel configuration
get_reading takes a units argument, which changes the call to the
Lakeshore depending on what units you want your messurement in. However,
the channels are configured with a unit for readout via the INTYPE
command. This update defaults to using this configured unit if None is
provided to get_reading(), which avoids the acquisition process from
needing to know the units needed when collecting data.