HostMaster Agent, to manage host's Agent instances.
The HostMaster currently exposes a single Process, whose job is to
start up and maintain all of the Agent instances on a single host.
When the process ends, the instances are killed.  Currently it's one
size fits all... but future work can allow start/stop/restart of
individual Agents, and incorporate different policies for handling
restarts, shutdown protocol, etc.