Fix several failing LS372 unit tests
A couple of heater tests were failing, this code is still being developed, so
that's not surprising. The one surprising one was a failing test on setting the
channel excitation current. This was likely failing due to autoscan being on,
but also maybe due to the test set range being out of valid range for the given
channel. Now we enable autorange, give it a few seconds, then disable autorange
and try to set one higher. This should be safe, as the auto-range should bring
it to a nominal value (unlikely to be the max range), then we can go one higher
for the test.