OCSAgent/OpSession: fix double-stop bug.
OpSession.set_status(...) now prevents backwards movements in the life
cycle (such as from "done" to "stopping").  This fixes the bug noticed
in HostMaster agent.  Allowed status strings are now explicitly
documented.

ocsbow is updated to handle the protocol correctly and more verbosely
and more generally usefully.  client_wampy.request() propagates
keywork args, so that timeout=... can be passed to
Operation.wait... this is not the best interface so don't get too
attached to it.