Attempt 1st communication with 240 again before failing
We'd commonly run into an issue the first time an LS240 was connected to a
computer where the first attempt at communicating with it would timeout and
you'd have to kill the Agent and try again. This simply tries this initial
communication (which is just a query for the serial number) up to three times
before failing. This has successfully gotten around this particular timeout
issue.