Separate synchronous and asynchronous functionality.
This separates the sync and async functionality into the Aggregator
and AggregatorAgent classes. The AggregatorAgent provides the crossbar
interface, and is safe to use asynchronously. Now all the non-threadsafe
data management happens within the Aggregator class, which is instantiated
locally each time the `run` process is called, so hopefully there is no
confusion as to which functions can be run from what contexts.