Aggregator 2.0 (#78)
This is a squash and merge, this is the squashed commit history:

* Refactors aggregator and adds lots of features and good things.

* Adds close_file function to G3FileRotator

* Removed locking from provider objects and frame collection

All locking is now in the `incoming_data` queue, which is processed
every run cycle. No other data should be touched outside of the
run thread context...

* Separate synchronous and asynchronous functionality.

This separates the sync and async functionality into the Aggregator
and AggregatorAgent classes. The AggregatorAgent provides the crossbar
interface, and is safe to use asynchronously. Now all the non-threadsafe
data management happens within the Aggregator class, which is instantiated
locally each time the `run` process is called, so hopefully there is no
confusion as to which functions can be run from what contexts.

* set's session status for aggregator process.

* Updates AggregatorAgent docs.

* Moved data processing logic into aggregator run thread
and implemented other comments suggested by Matthew...