Dockerize and document ocs web interface (#89)
This adds a Dockerfile in ocs/www for the web interface which builds a simple
image based on the OCS image with nginx installed and running by default. This
will serve the www/ directory at the web root served by nginx.

I also document how to use/setup OCS Web, both directly on the system running
it (without a web server in front) and remotely, recommending ssh tunneling for
the crossbar connection for now.

Finally I make the default realm "test_realm" as that's the default we have
everyone configure to. This will ease the setup for users, rather than having
them change from "debug_realm" each time they connect.