Handle ClientError when writing to InfluxDB
At Penn we encountered a situation where an old version of the Bluefors log
agent initialized an InfluxDB series with string values. After we updated the
Bluefors Agent it was sending floats (as it should). InfluxDB doesn't handle
this, it expects strings, matching the type of previous values. This requires
user intervention. Here we just log the error for the user and move on to
publishing other data values. This prevents us from crashing completely and
will continue to publish valid data points to the DB.

Fixes #98