Set the locale within Docker
Recent installations of a python package, watchdog, were running into a problem
decoding some part of a changelog. It turns out this was caused by the setting
of the locale within the Ubuntu 18.04 base Docker image. This line sets the
LANG to a valid one, specifically C.UTF-8.

Information about setting the Locales is provided on the Ubuntu Docker Hub
page:
https://hub.docker.com/_/ubuntu/?tab=description