Verify message data arrays during Feed.publish_message()
The message contents of aggregated feeds should only contain values supported
by the .g3 data container. We have run into issues with strings accidentally
being passed. This adds a verification during the publish_message call that
would cause a ValueError to raise on the Agent end of things if a wrong data
type is passed. This will make it obvious to the user that they've tried to
send malformed data.

fixes #103