Incorporate select changes from #107
This incorporates several changes from #107, but with some modifications. Most
notably it includes format_timeline(), which comes with a dramatic performance
increase, both by publishing all fields that share a timeline simultaneously and
by utilizing InfluxDB's line protocol, which is faster than the previously used
json based protocol.

We maintain the use of a local sqlite DB instead of switching to using InfluxDB
to track published .g3 files. We also remove the use of an additional md5sum
tag on data uploaded with this script.

Many thanks to @ahincks for the work this was based on.