Fix parsing for files with no feeds (#156)
Older files might not have a "feeds.NAME" in the agent address, which breaks
some of the parsing here. This will now upload data without the feeds tags that
are now present.

This isn't a huge deal breaker, as most (all?) agents just use one feed at this
time. All past data will be returned if the feed tag is left off of InfluxDB
queries.