Create ocs-crossbar Dockerfile and remove sisock refs
"sisock-crossbar" is the old crossbar server provided by the sisock repo.
Per #135, we want to redraw the separation between sisock and ocs. This creates
a Dockerfile from which we build a crossbar server Docker image with a sensible
default ocs-centric crossbar configuration already in the image. We also rename
all references to "sisock-crossbar" to simply just "crossbar" in the docs.

closes #135