Allow some time for reconnection to realm
Add the realm_joined attribute, which tracks whether the Agent is connected to
the realm. This allows us to wait for a reconnection onDisconnect before
shutting down the reactor in case the crossbar server comes back quickly. Else
we want to shutdown properly and close the Agent by stopping the reactor.