Unset subscribed topics, forcing a resubscribe if rejoining
If the crossbar server gets reset then the subscribed to topics are still
tracked by the Agent, but no longer work, since its a new crossbar server.
Resetting the subscribed_topics when we leave the realm forces a resubscribe
when we rejoin the realm, whether on a new crossbar server or not.

Without this, we'd get Agents that reconnect to a new crossbar server, but
don't get any data on the feeds they thought they were still subscribed to.