Remove empty blocks' block_name from frame (#179)
* Remove empty blocks' block_name from frame

If a previously published block wasn't published during a frame, the block's
name would end up in the frame's "block_names" list anyway. We now remove it if
the block is empty, i.e. if there are no timestamps.

* Build block_names correctly the first time

Instead of removing names that shouldn't be in the list, build it right the
first time.

* Extend block_names if already in frame