Add Op params to OpSession attributes
This would allow tracking of which params were used at startup of an Op so that
it could be resumed if needed during the onLeave/onJoin. However, I realized
this shouldn't be needed, as the process should keep running and we can close
it down safely in onDisconnect before stopping the reactor instead of doing it
onLeave.