Clarify the permissions structure for ocs-util
Since this requires creating the ocs user, we move that information from the
Aggregator Agent page to the more general (and visible) location in the
installation instructions.

We also set and expose the same port as is recommended for use in the ocs-util
instructions. This should help users who might run the container directly,
rather than through docker-compose, else they might run into the situation
where they try to access Jupyter via the recommended port 8880, but the service
is running on the default 8888.