HostMaster: manage services defined in docker-compose
Gives HostMaster the ability to scan docker-compose files, determine
whether each service is running, and bring services up and down as
though they were agents.

The mapping is imperfect -- HostMaster only knows the services names
and not the instance-id or agent-class.  But you can monitor the
state, and request up/down.