HostManager: assertively manage Docker-based Agents
HostManager expects to be the only one starting/stopping those
containers.  Also HostManager will shut everything down when it exits
due to SIGINT, as with systemctl stop.

Before this change, it was tolerated for docker-based agents to be
brought up outside the HostManager, i.e. directly by a user running
docker-compose up.  We can make this more configurable on a per-agent
or per-manager basis, if need be.