Try unsubbing onJoin before resubbing as usual
I think this does solve the problem (though haven't checked output data for
that), because it forces a new subscription ID and causes the old subscription
to become inactive (is_active=False). However, we are generating an "Unhandled
error in Deferred" that I can't seem to find and that has no traceback.